improve threading, style; add translate direction error detect; escape xml

+++ diy/mosesserver.py
@@ -7,14 +7,14 @@ import socket
 import signal
 import time
 from functools import lru_cache
+from operator import itemgetter
 import resource
 
 from zhconv import convert as zhconv
-from pangu import spacing
 import zhutil
 import jieba
 import jiebazhc
-from sqlitecache import SqliteCache
+from sqlitecache import LRUCache, SqliteCache
 from config import *
 
 #_curpath = os.path.normpath(os.path.join(os.getcwd(), os.path.dirname(__file__)))
@@ -23,8 +23,6 @@ from config import *
 os.chdir(os.path.dirname(os.path.abspath(sys.argv[0])))
 
 SIGNUM2NAME = dict((k, v) for v, k in signal.__dict__.items() if v.startswith('SIG') and not v.startswith('SIG_'))
-
-cache = SqliteCache(DB_zhccache, DB_zhccache_maxlen)
 
 resource.setrlimit(resource.RLIMIT_RSS, (MOSES_MAXMEM*1024 - 10000, MOSES_MAXMEM*1024))
 
@@ -39,88 +37,125 @@ notwhite = lambda x: x not in whitespace
 
 RE_WS_IN_FW = re.compile(r'([\u2018\u2019\u201c\u201d\u2e80-\u312f\u3200-\u32ff\u3400-\u4dbf\u4e00-\u9fff\uf900-\ufaff\uff00-\uffef])\s+(?=[\u2018\u2019\u201c\u201d\u2e80-\u312f\u3200-\u32ff\u3400-\u4dbf\u4e00-\u9fff\uf900-\ufaff\uff00-\uffef])')
 
-detokenize = lambda s: spacing(RE_WS_IN_FW.sub(r'\1', s)).strip()
+detokenize = lambda s: RE_WS_IN_FW.sub(r'\1', xml_unescape(s)).strip()
 
 quiet = False
 verbose = False
 
 runmoses = lambda mode: subprocess.Popen(m2c, shell=False, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=sys.stderr, cwd=MOSES_CWD) if mode=='m2c' else subprocess.Popen(c2m, shell=False, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=sys.stderr, cwd=MOSES_CWD)
 
-jieba.initialize(DICT_SMALL)
-jiebazhc.initialize()
+xml_escape_table = {
+	"&": "&amp;", '"': "&quot;", "'": "&apos;",
+	">": "&gt;", "<": "&lt;",
+}
 
-class MosesManager:
+xml_unescape_table = {
+	"&amp;": "&", "&quot;": '"', "&apos;": "'",
+	"&gt;": ">", "&lt;": "<",
+}
+
+def xml_escape(text):
+    """Produce entities within text."""
+    return "".join(xml_escape_table.get(c,c) for c in text)
+
+def xml_unescape(text):
+    """Produce entities within text."""
+    return " ".join(xml_unescape_table.get(c,c) for c in text.split(' '))
+
+class MosesManagerThread:
 	def __init__(self):
+		self.sqlcache = SqliteCache(DB_zhccache, DB_zhccache_maxlen)
+		self.lrucache = LRUCache(128)
 		self.pc2m = runmoses('c2m')
 		self.pm2c = runmoses('m2c')
 		sys.stderr.write('Started Moses c2m: %s\n' % self.pc2m.pid)
 		sys.stderr.write('Started Moses m2c: %s\n' % self.pm2c.pid)
 		sys.stderr.write('System ready.\n')
 		sys.stderr.flush()
-		self.lastupdatetime = 0
-		self.timediff = 3600
+		self.taskqueue = []
+		self.resultqueue = []
 
-	@lru_cache(maxsize=64)
-	def translatesentence(self, s, mode):
-		if time.time() - self.lastupdatetime > self.timediff:
-			sys.stderr.write(time.strftime("# %Y-%m-%d %H:%M:%S\n", time.gmtime()))
-			self.lastupdatetime = time.time()
-		for j in range(5):
-			if mode == "c2m":
-				rv = cache.get(s)
-				if rv:
-					return (rv, 0)
-				returncode = self.pc2m.poll()
-				if returncode is not None:
-					sys.stderr.write('Moses c2m (%s) is dead: %s\n' % (self.pc2m.pid, SIGNUM2NAME.get(-returncode, str(returncode))))
-					self.pc2m.wait()
-					self.pc2m = runmoses('c2m')
-					sys.stderr.write('Restarted Moses c2m: %s\n' % self.pc2m.pid)
-				proc = self.pc2m
-				tok = ' '.join(zhutil.addwalls(filter(notwhite, jiebazhc.cut(s,cut_all=False))))
-				#print(tok)
-			else:
-				returncode = self.pm2c.poll()
-				if returncode is not None:
-					sys.stderr.write('Moses m2c (%s) is dead: %s\n' % (self.pm2c.pid, SIGNUM2NAME.get(-returncode, str(returncode))))
-					self.pm2c.wait()
-					self.pm2c = runmoses('m2c')
-					sys.stderr.write('Restarted Moses m2c: %s\n' % self.pm2c.pid)
-				proc = self.pm2c
-				tok = ' '.join(zhutil.addwalls(filter(notwhite, jieba.cut(s,cut_all=False))))
-			proc.stdin.write(('%s\n' % tok).encode('utf8'))
-			proc.stdin.flush()
-			rv = detokenize(proc.stdout.readline().decode('utf8'))
-			if rv:
-				break
+	def checkmoses(self, mode):
 		if mode == "c2m":
-			cache.add(s, rv)
-			if time.time() - self.lastupdatetime > self.timediff:
-				cache.gc()
-		self.lastupdatetime = time.time()
-		return (rv, 1)
+			returncode = self.pc2m.poll()
+			if returncode is not None:
+				sys.stderr.write('Moses c2m (%s) is dead: %s\n' % (self.pc2m.pid, SIGNUM2NAME.get(-returncode, str(returncode))))
+				self.pc2m.wait()
+				self.pc2m = runmoses('c2m')
+				sys.stderr.write('Restarted Moses c2m: %s\n' % self.pc2m.pid)
+			return self.pc2m
+		else:
+			returncode = self.pm2c.poll()
+			if returncode is not None:
+				sys.stderr.write('Moses m2c (%s) is dead: %s\n' % (self.pm2c.pid, SIGNUM2NAME.get(-returncode, str(returncode))))
+				self.pm2c.wait()
+				self.pm2c = runmoses('m2c')
+				sys.stderr.write('Restarted Moses m2c: %s\n' % self.pm2c.pid)
+			return self.pm2c
+
+	def tokenize(self, s, mode):
+		if mode == "c2m":
+			return ' '.join(zhutil.addwalls(map(xml_escape, filter(notwhite, jiebazhc.cut(s)))))
+		else:
+			return ' '.join(zhutil.addwalls(map(xml_escape, filter(notwhite, jieba.cut(s)))))
+
+	def processtasks(self, mode):
+		if not self.taskqueue:
+			return True
+		for sn, s in self.taskqueue:
+			proc = self.checkmoses(mode)
+			tok = self.tokenize(s, mode)
+			proc.stdin.write(('%s\n' % tok).encode('utf8'))
+		proc.stdin.flush()
+		for sn, s in self.taskqueue:
+			rv = detokenize(proc.stdout.readline().decode('utf8'))
+			if not rv:
+				return False
+			self.resultqueue.append((sn, rv))
+			self.lrucache.add((s, mode), rv)
+			if mode == "c2m":
+				self.sqlcache.add(s, rv)
+		return True
+
+	def getcache(self, text, mode):
+		rv = self.lrucache.get((text, mode))
+		if rv is None and mode == "c2m":
+			rv = self.sqlcache.get(text)
+		return rv
 
 	def translate(self, text, mode, withcount=False):
-		outputtext = []
-		nohitcount = 0
+		hitcount = 0
+		misscount = 0
+		snum = 0
+		self.taskqueue = []
+		self.resultqueue = []
 		for l in text.split('\n'):
-			outputtext.append('<p>')
-			sentences = zhutil.splitsentence(zhconv(l.strip(), 'zh-cn'))
+			sentences = zhutil.splithard(zhconv(l.strip(), 'zh-cn'), 128)
 			for s in sentences:
-				res = ''
-				# prevent mem explode
-				for i in range(0, len(s), 128):
-					rv, nohit = self.translatesentence(s[i:i+128], mode)
-					res += rv
-					nohitcount += nohit
-				outputtext.append(res)
-			outputtext.append('</p>\n')
+				if s:
+					crv = self.getcache(s, mode)
+					if crv:
+						self.resultqueue.append((snum, crv))
+						hitcount += 1
+					else:
+						self.taskqueue.append((snum, s))
+						misscount += 1
+				else:
+					self.resultqueue.append((snum, s))
+				snum += 1
+			self.resultqueue.append((snum, '\n'))
+			snum += 1
+		self.processtasks(mode) # Error handling?
+		self.resultqueue.sort()
+		sys.stderr.write('%s Translated %s/%s sentences, %s.\n' % (time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime()), misscount, hitcount + misscount, mode))
+		ig = itemgetter(1)
+		outputtext = ''.join(map(ig, self.resultqueue))
 		if withcount:
-			return (''.join(outputtext), nohit)
+			return (outputtext, misscount)
 		else:
-			return ''.join(outputtext)
+			return outputtext
 
-def handle(data):
+def handle(mm, data):
 	oper = umsgpack.loads(data)
 	if oper[0] == 'c2m':
 		return umsgpack.dumps(mm.translate(oper[1], 'c2m', oper[2]))
@@ -166,9 +201,11 @@ def receive(filename, data):
 	return received
 
 def serve(filename):
+	mm = None
 	if os.path.exists(filename):
 		try:
 			receive(filename, umsgpack.dumps(('ping',)))
+			print("Server already started.")
 			return False
 		except Exception:
 			# not removed socket
@@ -178,11 +215,13 @@ def serve(filename):
 		with socket.socket(socket.AF_UNIX, socket.SOCK_STREAM) as sock:
 			sock.bind(filename)
 			sock.listen(5)
-			jieba.initialize()
+			jieba.initialize(DICT_SMALL)
+			jiebazhc.initialize()
+			mm = MosesManagerThread()
 			while 1:
 				conn, addr = sock.accept()
 				received = recvall(conn, 1024)
-				result = handle(received)
+				result = handle(mm, received)
 				if result is None:
 					conn.sendall(b"\xc0")
 				elif result == b'stop':
@@ -193,9 +232,10 @@ def serve(filename):
 					conn.sendall(result)
 				conn.close()
 	finally:
-		mm.pc2m.terminate()
-		mm.pm2c.terminate()
-		cache.gc()
+		if mm:
+			mm.pc2m.terminate()
+			mm.pm2c.terminate()
+			mm.sqlcache.gc()
 		if os.path.exists(filename):
 			os.unlink(filename)
 		print("Server stopped.")
@@ -203,7 +243,6 @@ def serve(filename):
 if __name__ == '__main__':
 	filename = MS_SOCK
 	try:
-		mm = MosesManager()
 		serve(filename)
 	except OSError as ex:
 		if 'Address already in use' in str(ex):
